github-app/requirements.txt
fastapi>=0.115
uvicorn>=0.34
httpx>=0.28
PyGithub>=2.6
cryptography>=44.0
